requirements.txt
opik
groq
typer[all]
pydantic>=2
tqdm
rapidfuzz
python-dotenv
pytest


